+++ src/types.ts
@@ -1,5 +1,6 @@
 import type { Type } from "ts-morph";
 import type { FakerLocale } from "./constants";
+import type { Low } from "lowdb";
 
 export type EvaluatedFakerArgs = { path: string; args: any } | undefined;
 export type LazyFaker = typeof import("@faker-js/faker").faker;
@@ -19,10 +20,26 @@ export type FakerEngineOptions = {
   locale?: FakerLocale;
 };
 
+export type DatabaseOptions = {
+  enabled: boolean;
+  dest?: string;
+};
+
+type BrowserExposeOptions = {
+  name: string;
+  mode: "module" | "global";
+};
+
+export type BrowserOptions = {
+  expose?: Partial<BrowserExposeOptions>;
+};
+
 export type ConfigOptions = {
   sourcePath: string | string[];
   server?: ServerOptions;
   faker?: FakerEngineOptions;
+  database?: DatabaseOptions;
+  browser?: BrowserOptions;
 };
 
 export type UserConfig = {
@@ -37,8 +54,15 @@ export type ForgeOptions = {
   count?: string;
 };
 
+export type Entity = {
+  type: Type;
+  filepath: string;
+  tablepath: string;
+  table: Low<unknown[]>;
+};
+
 export interface IGenerated {
-  readonly entities: Map<string, { type: Type; filepath: string }>;
+  readonly entities: Map<string, Entity>;
   forge: (type: Type, options: ForgeOptions) => Promise<GeneratorForge>;
 }